rename uptime_monitor_checks
