+++ snuba/datasets/processors/uptime_monitors_processor.py
@@ -6,9 +6,9 @@ from snuba.utils.metrics.wrapper import MetricsWrapper
 
 logger = logging.getLogger(__name__)
 
-metrics = MetricsWrapper(environment.metrics, "uptime_monitors.processor")
+metrics = MetricsWrapper(environment.metrics, "uptime_monitor_checks.processor")
 
 
-class UptimeMonitorsProcessor(RustCompatProcessor):
+class UptimeMonitorChecksProcessor(RustCompatProcessor):
     def __init__(self) -> None:
-        super().__init__("UptimeMonitorsProcessor")
+        super().__init__("UptimeMonitorChecksProcessor")
