+++ snuba/settings/__init__.py
@@ -116,7 +116,7 @@ CLUSTERS: Sequence[Mapping[str, Any]] = [
             "group_attributes",
             "generic_metrics_gauges",
             "profile_chunks",
-            "uptime_monitors",
+            "uptime_monitor_checks",
         },
         "single_node": True,
     },
