+++ snuba/settings/settings_test_distributed_migrations.py
@@ -59,7 +59,7 @@ CLUSTERS: Sequence[Mapping[str, Any]] = [
             "group_attributes",
             "generic_metrics_gauges",
             "profile_chunks",
-            "uptime_monitors",
+            "uptime_monitor_checks",
         },
         "single_node": False,
         "cluster_name": "storage_cluster",
