+++ snuba/utils/streams/topics.py
@@ -45,7 +45,7 @@ class Topic(Enum):
     PROFILE_CHUNKS = "snuba-profile-chunks"
 
     REPLAYEVENTS = "ingest-replay-events"
-    UPTIME_MONITORS = "snuba-uptime-monitors-events"
+    UPTIME_MONITOR_CHECKS = "snuba-uptime-monitor-checks"
     GENERIC_METRICS = "snuba-generic-metrics"
     GENERIC_METRICS_SETS_COMMIT_LOG = "snuba-generic-metrics-sets-commit-log"
     GENERIC_METRICS_DISTRIBUTIONS_COMMIT_LOG = (
@@ -60,7 +60,7 @@ class Topic(Enum):
     DEAD_LETTER_METRICS = "snuba-dead-letter-metrics"
     DEAD_LETTER_GENERIC_METRICS = "snuba-dead-letter-generic-metrics"
     DEAD_LETTER_REPLAYS = "snuba-dead-letter-replays"
-    DEAD_LETTER_UPTIME_MONITORS = "snuba-dead-letter-uptime-monitors"
+    DEAD_LETTER_UPTIME_MONITOR_CHECKS = "snuba-dead-letter-uptime-monitor-checks"
     DEAD_LETTER_GENERIC_EVENTS = "snuba-dead-letter-generic-events"
     DEAD_LETTER_QUERYLOG = "snuba-dead-letter-querylog"
     DEAD_LETTER_GROUP_ATTRIBUTES = "snuba-dead-letter-group-attributes"
